Adjustments for types. (#178)
The internalApi starts as undefined, gets set to an object when connecting to the ngrok process and when the process is killed is set to null. I wanted to simplify the states here to only either null or the request object. Consequently the return type of getApi should also be updated to reflect that it may be null. If you'd prefer not to change the states, I could change the return type of getApi to RequestAPI<Request, CoreOptions, RequiredUriUrl> | null | undefined.

Whilst making the changes for internalApi, it seemed more consistent to do the same for processUrl and the getUrl function.

What do you think?