Tidy up errors when starting tunnels.
When we try to start a tunnel it will either return a public url or throw an NgrokClientError. We don't need to turn the NgrokClientError into a new error if it isn't retriable, the NgrokClientError has all the details one could want. This cleans up the connectRetry function and returns a more consistent error in the case where we can't start a tunnel.