fix: handles things if the log format is set to json (#221)
Previously, if the log_format in the config is set to json, ngrok would start correctly but getProcess would never resolve. It wasn't resolving because it couldn't parse the API URL from the JSON string instead of the logfmt string. This fixes that, parsing the lines as JSON if they start with a curly brace. Fixes #210