Throw client errors correctly.
If the HTTP client catches an error but it doesn't have a response, then it is not an HTTP error, so we'll throw the original error instead. This will surface what is causing an error that isn't because of an HTTP response in the client.

Should fix #229, or at least help diagnose it.